Center_point_Cartesian_coordinated_system
You will be given the coordinates of two points on a Cartesian coordinate system - X1, Y1, X2, and Y2 on separate lines. Write a function that prints the point which is closest to the center of the coordinate system (0, 0) in the format: "({X}, {Y})"
If the points are at the same distance from the center, print only the first one. The resulting coordinates must be formatted to the lower integer.